Address some more commentS

+++ crates/provider/src/shared/constants.rs
@@ -3,4 +3,4 @@ pub const DEFAULT_REMOTE_DIR: &str = "/cfg";
 pub const DEFAULT_DATA_DIR: &str = "/data";
 pub const LOCALHOST: &str = "127.0.0.1";
 /// The port substrate listens for p2p connections on
-pub const P2P_PORT: u32 = 30333;
+pub const P2P_PORT: u16 = 30333;
